fix(stringify): handle multiple paragraphs in a list item
