+++ packages/comark-ansi/src/handlers/li.ts
@@ -1,5 +1,10 @@
 import type { NodeHandler } from 'comark/render'
 import type { ComarkElement, ComarkNode } from 'comark'
+import { indent } from 'comark/utils'
+
+// Block elements that need explicit indentation in list items.
+// Note: ol/ul are handled by their own handlers which manage indentation via listIndent context.
+const blockElements = new Set(['pre', 'blockquote', 'table'])
 
 export const li: NodeHandler = async (node, state) => {
   const children = node.slice(2) as ComarkNode[]
@@ -14,15 +19,31 @@ export const li: NodeHandler = async (node, state) => {
     prefix += input[1].checked || input[1][':checked'] ? '[x] ' : '[ ] '
   }
 
-  let content = ''
+  const prefixWidth = prefix.length
+  let result = ''
   for (const child of children) {
-    content += await state.one(child, state, node)
+    const rendered = await state.one(child, state, node)
+    if (result && Array.isArray(child)) {
+      if (blockElements.has(child[0] as string)) {
+        // Block-level child: put on its own line and indent to align with list prefix
+        const indented = indent(rendered, { width: prefixWidth })
+        result = result.trimEnd() + '\n' + indented.trimEnd() + '\n'
+        continue
+      }
+
+      if (child[0] === 'p') {
+        const indented = indent(rendered, { width: prefixWidth })
+        result = result.trimEnd() + '\n\n' + indented.trimEnd() + '\n'
+        continue
+      }
+    }
+    result += rendered
   }
-  content = content.trim()
+  result = result.trim()
 
   if (typeof order === 'number') {
     state.applyContext({ order: order + 1 })
   }
 
-  return `${prefix}${content}\n`
+  return `${prefix}${result}\n`
 }
