+++ packages/comark-ansi/src/handlers/p.ts
@@ -1,11 +1,16 @@
 import type { NodeHandler } from 'comark/render'
 import type { ComarkNode } from 'comark'
 
-export const p: NodeHandler = async (node, state) => {
+export const p: NodeHandler = async (node, state, parent) => {
   const children = node.slice(2) as ComarkNode[]
   let result = ''
   for (const child of children) {
     result += await state.one(child, state, node)
   }
+
+  if (parent?.[0] === 'li') {
+    return result
+  }
+
   return result + '\n\n'
 }
