+++ packages/comark/src/internal/stringify/handlers/mdc.ts
@@ -1,6 +1,6 @@
 import type { State } from 'comark/render'
 import type { ComarkElement, ComarkNode } from 'comark'
-import { indent } from '../indent.ts'
+import { indent } from '../../../utils/index.ts'
 import { comarkAttributes, comarkYamlAttributes } from '../attributes.ts'
 import { html } from './html.ts'
