+++ packages/comark/src/internal/stringify/handlers/ol.ts
@@ -1,6 +1,6 @@
 import type { State } from 'comark/render'
 import type { ComarkElement, ComarkNode } from 'comark'
-import { indent } from '../indent.ts'
+import { indent } from '../../../utils/index.ts'
 
 export async function ol(node: ComarkElement, state: State) {
   const children = node.slice(2) as ComarkNode[]
